+++ src/index.js
@@ -51,7 +51,6 @@ import { App as SlackApp } from './utils/slack/bolt.cjs';
 import ConfigurationController from './controllers/configuration.js';
 import FulfillmentController from './controllers/event/fulfillment.js';
 import ImportController from './controllers/import.js';
-import AssistantController from './controllers/assistant.js';
 import { s3ClientWrapper } from './support/s3.js';
 import { multipartFormData } from './support/multipart-form-data.js';
 import ApiKeyController from './controllers/api-key.js';
@@ -100,11 +99,10 @@ async function run(request, context) {
       trigger,
       FulfillmentController(context),
       ImportController(context),
-      AssistantController(context),
       ApiKeyController(context),
       SitesAuditsToggleController(context.dataAccess),
       OpportunitiesController(context.dataAccess),
-      SuggestionsController(context.dataAccess),
+      SuggestionsController(context.dataAccess, context.sqs, context.env),
     );
 
     const routeMatch = matchPath(method, suffix, routeHandlers);