+++ CHANGELOG.md
@@ -1,3 +1,101 @@
+# [1.98.0](https://github.com/adobe/spacecat-api-service/compare/v1.97.1...v1.98.0) (2025-03-05)
+
+
+### Features
+
+* introduce Bulk scrape and audit ([#806](https://github.com/adobe/spacecat-api-service/issues/806)) ([671a7aa](https://github.com/adobe/spacecat-api-service/commit/671a7aa6d7c6dad9dd1288e7c5e0355ab501f546))
+
+## [1.97.1](https://github.com/adobe/spacecat-api-service/compare/v1.97.0...v1.97.1) (2025-03-04)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#804](https://github.com/adobe/spacecat-api-service/issues/804)) ([58d6f01](https://github.com/adobe/spacecat-api-service/commit/58d6f018cdd12f60f8fb9170214762d45c7caba3))
+
+# [1.97.0](https://github.com/adobe/spacecat-api-service/compare/v1.96.4...v1.97.0) (2025-03-04)
+
+
+### Features
+
+* **sites:** add delivery configuration object to site object ([#744](https://github.com/adobe/spacecat-api-service/issues/744)) ([a87c751](https://github.com/adobe/spacecat-api-service/commit/a87c7518f9950d0f5f6379eed406946727c1c764))
+
+## [1.96.4](https://github.com/adobe/spacecat-api-service/compare/v1.96.3...v1.96.4) (2025-03-04)
+
+
+### Bug Fixes
+
+* adding all-traffic to import schemas ([#803](https://github.com/adobe/spacecat-api-service/issues/803)) ([980dbef](https://github.com/adobe/spacecat-api-service/commit/980dbef464ff7e99cc34295a38bde66043d34c97))
+
+## [1.96.3](https://github.com/adobe/spacecat-api-service/compare/v1.96.2...v1.96.3) (2025-03-04)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#802](https://github.com/adobe/spacecat-api-service/issues/802)) ([f6a25f5](https://github.com/adobe/spacecat-api-service/commit/f6a25f502736afdceadd3319c274776134418c2c))
+
+## [1.96.2](https://github.com/adobe/spacecat-api-service/compare/v1.96.1...v1.96.2) (2025-03-04)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#801](https://github.com/adobe/spacecat-api-service/issues/801)) ([461e31e](https://github.com/adobe/spacecat-api-service/commit/461e31e9a77b602777cd48c59ab4ca661fe43f15))
+
+## [1.96.1](https://github.com/adobe/spacecat-api-service/compare/v1.96.0...v1.96.1) (2025-03-03)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#799](https://github.com/adobe/spacecat-api-service/issues/799)) ([912646d](https://github.com/adobe/spacecat-api-service/commit/912646d64cfc62233d5dbca1f216f7ef5e1d867e))
+
+# [1.96.0](https://github.com/adobe/spacecat-api-service/compare/v1.95.0...v1.96.0) (2025-03-03)
+
+
+### Features
+
+* **bulk-audit-enable-disable:** bulk enable/disable audits with csv support ([#797](https://github.com/adobe/spacecat-api-service/issues/797)) ([2ecd0cf](https://github.com/adobe/spacecat-api-service/commit/2ecd0cfe6f75061b03cc088752fc582f7ba10a15))
+
+# [1.95.0](https://github.com/adobe/spacecat-api-service/compare/v1.94.1...v1.95.0) (2025-03-03)
+
+
+### Features
+
+* add autofix endpoint ([#773](https://github.com/adobe/spacecat-api-service/issues/773)) ([53ab57f](https://github.com/adobe/spacecat-api-service/commit/53ab57f623c355ac9a20ff2dc5ed20d6cce84768))
+
+## [1.94.1](https://github.com/adobe/spacecat-api-service/compare/v1.94.0...v1.94.1) (2025-03-02)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#798](https://github.com/adobe/spacecat-api-service/issues/798)) ([07a57a7](https://github.com/adobe/spacecat-api-service/commit/07a57a792933fa7d0cde4b1349deb21e1e989011))
+
+# [1.94.0](https://github.com/adobe/spacecat-api-service/compare/v1.93.20...v1.94.0) (2025-02-28)
+
+
+### Features
+
+* introduce google authentication status endpoint ([#795](https://github.com/adobe/spacecat-api-service/issues/795)) ([d09d537](https://github.com/adobe/spacecat-api-service/commit/d09d537e2d146d024dc5b380fbf86dce9a084a01))
+
+## [1.93.20](https://github.com/adobe/spacecat-api-service/compare/v1.93.19...v1.93.20) (2025-02-28)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.9.5 ([#796](https://github.com/adobe/spacecat-api-service/issues/796)) ([38bed4e](https://github.com/adobe/spacecat-api-service/commit/38bed4ef13191c10be92ab3f89a6a7985c56de27))
+
+## [1.93.19](https://github.com/adobe/spacecat-api-service/compare/v1.93.18...v1.93.19) (2025-02-27)
+
+
+### Bug Fixes
+
+* Sites-29600 Remove Assistant's handlebar dependency ([#794](https://github.com/adobe/spacecat-api-service/issues/794)) ([ebdaf75](https://github.com/adobe/spacecat-api-service/commit/ebdaf75a4a60524c5240599eada3f0110083bd62))
+
+## [1.93.18](https://github.com/adobe/spacecat-api-service/compare/v1.93.17...v1.93.18) (2025-02-27)
+
+
+### Bug Fixes
+
+* sites-29600 remove import assistant endpoint and code ([#793](https://github.com/adobe/spacecat-api-service/issues/793)) ([186bb07](https://github.com/adobe/spacecat-api-service/commit/186bb078e094a4b4147f6945dcc43f244b58b1a0))
+
 ## [1.93.17](https://github.com/adobe/spacecat-api-service/compare/v1.93.16...v1.93.17) (2025-02-27)